+++ packages/core/src/docs/examples/row-grouping.stories.tsx
@@ -9,7 +9,8 @@ import {
 } from "../../data-editor/stories/utils.js";
 import { SimpleThemeWrapper } from "../../stories/story-utils.js";
 import { GridCellKind, type Item } from "../../internal/data-grid/data-grid-types.js";
-import { type RowGroupingOptions, useRowGrouping } from "../../data-editor/row-grouping.js";
+import { type RowGroupingOptions } from "../../data-editor/row-grouping.js";
+import { useRowGrouping } from "../../data-editor/row-grouping-api.js";
 
 export default {
     title: "Glide-Data-Grid/DataEditor Demos",
